requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-PyMongo==2.3.0
Flask-WTF==0.14.3
python-dotenv==0.19.2
Werkzeug==1.0.1